+++ src/iodata.cpp
@@ -183,13 +183,13 @@ chemtools::IOData chemtools::get_molecular_basis_from_fchk(const std::string& fc
   py::array_t<double, py::array::c_style> coords_atoms = iodata_obj.attr("atcoords").cast<py::array_t<double>>();
 
   // Get Charges of the atoms.
-  int* charges = new int[natoms];
-  py::array_t<int> charges_atoms = iodata_obj.attr("atcorenums").cast<py::array_t<int>>();
+  long int* charges = new long int[natoms];
+  py::array_t<long int> charges_atoms = iodata_obj.attr("atcorenums").cast<py::array_t<long int>>();
   for(int i = 0; i < natoms; i++) {charges[i] = charges_atoms.at(i);}
 
   // Get Atomic Numbers of the atoms
-  int* atnums = new int[natoms];
-  py::array_t<int> atnums_atoms = iodata_obj.attr("atnums").cast<py::array_t<int>>();
+  long int* atnums = new long int[natoms];
+  py::array_t<long int> atnums_atoms = iodata_obj.attr("atnums").cast<py::array_t<long int>>();
   for(int i = 0; i < natoms; i++) {atnums[i] = atnums_atoms.at(i);}
 
   // Commit them to memory so that pybind11 doesn't wipe them out. This is done using column order.
@@ -236,10 +236,10 @@ chemtools::IOData::IOData(const chemtools::IOData& copy):
 {
   int nbasis = copy.orbital_basis_.numb_basis_functions();
   orbital_basis_ = chemtools::MolecularBasis(copy.orbital_basis_);
-  charges_ = new int[natoms];
-  std::memcpy(charges_, copy.charges_, sizeof(int) * natoms);
-  atnums_ = new int[natoms];
-  std::memcpy(atnums_, copy.atnums_, sizeof(int) * natoms);
+  charges_ = new long int[natoms];
+  std::memcpy(charges_, copy.charges_, sizeof(long int) * natoms);
+  atnums_ = new long int[natoms];
+  std::memcpy(atnums_, copy.atnums_, sizeof(long int) * natoms);
   coord_atoms_ = new double[3 * natoms];
   std::memcpy(coord_atoms_, copy.coord_atoms_, sizeof(double) * 3 * natoms);
   one_rdm_ = new double[copy.one_rdm_shape_[0] * copy.one_rdm_shape_[0]];
